Change the header and remove irrelavent componenets
