+++ donor-portal/src/App.tsx
@@ -1,22 +1,18 @@
 import React, { useEffect, useState } from "react";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 import { useAuthContext } from "@asgardeo/auth-react";
-import FullBleedContainer from "./pages/layout/full-bleed-container";
 import NavBar from "./components/navbar";
 import Footer from "./components/footer";
 import Home from "./pages/home";
 import AboutUs from "./pages/about-us";
-import Donors from "./pages/donors";
 import Suppliers from "./pages/suppliers";
-import DonateNow from "./pages/donate-now";
 import "./App.css";
 import AidPackageDetailsPage from "./pages/package";
 import NewsRoom from "./pages/newsroom";
 import UserContext from "./userContext";
 import Http from "./apis/httpCommon";
 import AidPackageService from "./apis/services/AidPackageService";
-import IndividualDonors from "./pages/donors/individual-donors";
-import MedicalNeeds from "./pages/medical-needs";
+import DonateNowPage from "./pages/donate-now";
 
 function App() {
   const { state, httpRequest, getDecodedIDToken } = useAuthContext();
@@ -53,15 +49,6 @@ function App() {
             <Route exact path="/about-us">
               <AboutUs />
             </Route>
-            <Route exact path="/donors">
-              <Donors />
-            </Route>
-            <Route exact path="/donors/individual">
-              <IndividualDonors />
-            </Route>
-            <Route exact path="/medical-needs">
-              <MedicalNeeds />
-            </Route>
             <Route exact path="/suppliers">
               <Suppliers />
             </Route>
@@ -69,15 +56,13 @@ function App() {
               <AidPackageDetailsPage />
             </Route>
             <Route exact path="/donate-now">
-              <DonateNow />
+              <DonateNowPage />
             </Route>
             <Route exact path="/news-room">
               <NewsRoom />
             </Route>
           </Switch>
-          <FullBleedContainer>
-            <Footer />
-          </FullBleedContainer>
+          <Footer />
         </BrowserRouter>
       </UserContext.Provider>
     </div>
