+++ donor-portal/src/components/footer/styles.css
@@ -49,12 +49,12 @@
   margin-bottom: 30px;
 }
 
-.icons {
+.footer .icons {
   display: flex;
   font-size: 17px;
 }
 
-.icon {
+.footer .icon {
   color: #333333;
   align-self: center;
   margin: 10px;
@@ -63,7 +63,7 @@
   padding: 10%;
 }
 
-.icons p {
+.footer .icons p {
   letter-spacing: 3px;
   color: var(--elixir-color-gray-600);
 }
@@ -74,7 +74,7 @@
   text-decoration: none;
   font-size: 0.8125rem;
   margin: auto;
-  width: 65%;  
+  width: 65%;
 }
 
 .footer-gray-middle span {
@@ -83,7 +83,7 @@
 
 .text-link {
   color: white;
-  padding: 0 10px; 
+  padding: 0 10px;
 }
 
 .text-link:not(:last-child) {
@@ -107,7 +107,7 @@
 
 .footer-gray-bottom  a {
   text-decoration: underline;
-  color: #c3c3c3; 
+  color: #c3c3c3;
 }
 
 .footer-gray-bottom .nav {
