+++ donor-portal/src/components/navbar/index.tsx
@@ -15,47 +15,46 @@ export default function NavBar() {
   };
 
   return (
-    <nav className="navbar">
-      <Link to="/">
-        <div className="logo" />
-      </Link>
-      <div className="nav-links">
-        <NavLink to="/" className="nav-link">
-          Home
-        </NavLink>
-        <NavLink to="/about-us" className="nav-link">
-          About Us
-        </NavLink>
-        <NavLink exact to="/donors" className="nav-link">
-          How to Donate
-        </NavLink>
-        <NavLink to="/medical-needs" className="nav-link">
-          Medical Needs
-        </NavLink>
-        <NavLink to="/suppliers" className="nav-link">
-          Hospitals & Suppliers
-        </NavLink>
-        <NavLink to="/news-room" className="nav-link">
-          Newsroom
-        </NavLink>
-        {state.isAuthenticated ? (
-          <button
-            className="logout-btn"
-            type="button"
-            onClick={() => handleLogout()}
-          >
-            Logout
-          </button>
-        ) : (
-          <button
-            className="login-btn"
-            type="button"
-            onClick={() => handleLogin()}
-          >
-            Login
-          </button>
-        )}
-      </div>
-    </nav>
+    <div className="navbar">
+      <nav className="navbar-container">
+        <Link to="/">
+          <div className="logo" />
+        </Link>
+        <div className="nav-links">
+          <NavLink to="/" exact className="nav-link">
+            Home
+          </NavLink>
+          <NavLink to="/about-us" className="nav-link">
+            About Us
+          </NavLink>
+          <NavLink to="/donate-now" className="nav-link">
+            Donate Now
+          </NavLink>
+          <NavLink to="/suppliers" className="nav-link">
+            Hospitals & Suppliers
+          </NavLink>
+          <NavLink to="/news-room" className="nav-link">
+            Newsroom
+          </NavLink>
+          {state.isAuthenticated ? (
+            <button
+              className="login-btn nav-link"
+              type="button"
+              onClick={() => handleLogout()}
+            >
+              Logout
+            </button>
+          ) : (
+            <button
+              type="button"
+              className="login-btn nav-link"
+              onClick={() => handleLogin()}
+            >
+              Login
+            </button>
+          )}
+        </div>
+      </nav>
+    </div>
   );
 }
