+++ donor-portal/src/components/navbar/styles.css
@@ -1,228 +1,89 @@
-:root {
-  --elixir-navbar-background-color: #fff;
-  --elixir-navbar-shadow: var(--elixir-global-shadow-lw);
-  --elixir-navbar-height: 3.75rem;
-  --elixir-navbar-padding-vertical: calc(var(--elixir-spacing-vertical) * 0.5);
-  --elixir-navbar-padding-horizontal: var(--elixir-spacing-horizontal);
-  --elixir-z-index-fixed: 4;
-  --elixir-navbar-link-color: var(--elixir-font-color-base);
-  --elixir-navbar-item-padding-vertical: 0.25rem;
-  --elixir-navbar-item-padding-horizontal: 0.75rem;
-}
-
-/* .navbar {
-  display: flex;
-  box-sizing: border-box;
-  background-color: var(--elixir-navbar-background-color);
-  box-shadow: var(--elixir-navbar-shadow);
-  height: var(--elixir-navbar-height);
-  padding: var(--elixir-navbar-padding-vertical)
-    var(--elixir-navbar-padding-horizontal);
-}
-
-.navbar--fixed-top {
-  position: -webkit-sticky;
-  position: sticky;
-  top: 0;
-  z-index: var(--elixir-z-index-fixed);
-}
-
-.navbar__inner {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  width: 100%;
-  box-sizing: border-box;
-}
-
-.navbar__items {
-  box-sizing: border-box;
-  align-items: center;
-  display: flex;
-  flex: 1;
-  min-width: 0;
-  height: 100%;
-}
-
-.navbar__items--right {
-  flex: 0 0 auto;
-  justify-content: flex-end;
-}
-
-.navbar__toggle {
-  display: none;
-  margin-right: 0.5rem;
-}
-
-.clean-btn {
-  background: none;
-  border: none;
-  color: inherit;
-  cursor: pointer;
-  font-family: inherit;
-  padding: 0;
-}
-
-.navbar__brand {
-  align-items: center;
-  display: flex;
-  margin-right: 1rem;
-  min-width: 0;
-  color: var(--elixir-navbar-link-color);
-}
-
-.navbar__logo {
-  flex: 0 0 auto;
-  height: 2rem;
-  margin-right: 0.5rem;
-}
-
-.navbar__logo img {
-  height: 100%;
-}
-
-.themedImage--light {
-  display: initial;
-}
-
-.navbar__title {
-  flex: 1 1 auto;
-  overflow: hidden;
-  white-space: nowrap;
-  text-overflow: ellipsis;
-}
-
-.navbar__item {
-  display: inline-block;
-  padding: var(--elixir-navbar-item-padding-vertical)
-    var(--elixir-navbar-item-padding-horizontal);
-}
-
-.navbar__link {
-  color: var(--elixir-navbar-link-color);
-}
-
-.iconExternalLink_gzkf {
-  margin-left: 0.3rem;
-  position: relative;
-  top: 1px;
-} */
-
 .navbar {
-  height: 70px;
-  background-color: var(--elixir-color-red);
-  display: flex;
-  justify-content: space-between;
+    background-color: var(--elixir-color-red);
+}
+
+.navbar-container {
+    height: 70px;
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+    max-width: var(--elixir-content-max-width);
+    margin: 0 auto;
 }
 
 .logo {
-  background-image: url("../../../public/assets/images/logo.png");
-  height: 70px;
-  background-color: white;
-  background-repeat: no-repeat;
-  background-size: 170px 70px;
-  background-position: center;
-  width: 250px;
-  cursor: pointer;
+    background-image: url("../../../public/assets/images/logo.png");
+    height: 70px;
+    background-color: white;
+    background-repeat: no-repeat;
+    background-size: 170px 70px;
+    background-position: center;
+    width: 250px;
+    cursor: pointer;
 }
 
 .nav-links {
-  background-color: var(--elixir-color-red);
-  height: 100%;
-  display: flex;
-  align-items: center;
-  color: white;
-  margin-right: 70px;
+    background-color: var(--elixir-color-red);
+    height: 100%;
+    display: flex;
+    align-items: center;
+    color: white;
 }
 
 .nav-link {
-  color: white;
-  padding: 0 15px;
-  font-size: 16px;
-  font-weight: 600;
-  letter-spacing: 2px;
-  display: flex;
-  align-items: center;
-  position: relative;
-
-  transition: all 0.5s;
-  z-index: 2;
+    color: white;
+    padding: 0 15px;
+    display: flex;
+    align-items: center;
+    position: relative;
+    z-index: 2;
 }
 
-.nav-link.active::before {
-  content: "";
-  width: 50%;
-  height: 1px;
-  position: absolute;
-  left: 50%;
-  transform: translate(-50%);
-  bottom: -5px;
-  background-color: #fff;
-  transition: all 0.5s;
-  animation: 0.5s appearNavLink;
+.nav-link.active{
+    font-weight: 600;
 }
 
-@keyframes appearNavLink {
-  from {
-    transform: translate(-100%);
-  }
-
-  50% {
-    transform: translate(-20%);
-  }
-}
-
-.nav-link:not(:last-of-type) {
-  border-right: 1px solid white;
-}
-
-.nav-link.active,
 .nav-link:hover {
-  color: var(--elixir-color-gray-400);
-  /* padding: 2rem; */
+    color: var(--elixir-color-gray-400);
 }
 
 .logout-btn {
-  background: none;
-  border: none;
-  padding: 0;
-  color: white;
-  padding: 0 15px;
-  font-size: 16px;
-  font-weight: 600;
-  letter-spacing: 2px;
-  cursor: pointer;
+    background: none;
+    border: none;
+    color: white;
+    padding: 0 15px;
+    font-size: 16px;
+    font-weight: 600;
+    letter-spacing: 2px;
+    cursor: pointer;
 }
 
 .logout-btn:hover {
-  color: var(--elixir-color-gray-400);
+    color: var(--elixir-color-gray-400);
 }
 
 .login-btn {
-  background-color: #f7f7f7;
-  background-image: linear-gradient(#f7f7f7, #e7e7e7);
-  border-radius: 50%;
-  box-shadow: 3px 3px 1px rgb(151, 151, 151), inset 0 2px 3px #fff;
-  color: var(--elixir-color-red);
-  display: block;
-  font-size: 20px;
-  font-weight: 700;
-  text-decoration: none;
-  position: relative;
-  top: 20%;
-  width: 80px;
-  height: 80px;
-  z-index: 1;
-  text-align: center;
-  line-height: 70px;
-  border: none;
-  cursor: pointer;
-  transition: all 0.5s;
-  margin-left: 8px;
+    background-color: #f7f7f7;
+    background-image: linear-gradient(#f7f7f7, #e7e7e7);
+    border-radius: 50%;
+    box-shadow: 3px 3px 1px rgb(190, 190, 190), inset 0 2px 3px #fff;
+    color: var(--elixir-color-red);
+    display: block;
+    font-size: 17px;
+    font-weight: 700;
+    text-decoration: none;
+    position: relative;
+    top: 20%;
+    width: 75px;
+    height: 75px;
+    z-index: 1;
+    text-align: center;
+    line-height: 70px;
+    border: none;
+    cursor: pointer;
+    margin-left: 8px;
 }
 
 .login-btn:hover {
-  color: red;
-  box-shadow: 5px 5px 1px rgb(151, 151, 151), inset 0 2px 3px #fff;
-  transform: translate(-3px, -3px);
+    color: red;
 }
