+++ donor-portal/src/pages/about-us/styles.css
@@ -43,11 +43,6 @@
   text-align: center;
 }
 
-.container {
-  display: flex;
-  margin-top: 2rem;
-}
-
 .leadership-section {
   margin-top: 2rem;
 }
