+++ donor-portal/src/pages/donate-now/styles.css
@@ -1,57 +1,83 @@
-.donate-now-page .header-image-content {
-  max-width: var(--elixir-content-max-width);
-  flex: 1;
+.medical-needs-page {
+  margin-top: 48px;
 }
 
-.donate-now-page .header-image h1 {
-  color: var(--elixir-color-white);
-  margin-left: 2rem;
-}
-
-.donate-now-container {
-  padding: 0 2rem;
-}
-
-.donation-org-container {
-  margin-top: 5rem;
-}
-
-.donation-org-group {
-  margin-bottom: 4rem;
-}
-
-.donation-org-row {
+.medical-needs-heading {
   display: flex;
+  justify-content: space-between;
   align-items: center;
-  gap: 1rem;
-  border-bottom: 2px solid var(--elixir-color-gray-500);
-  margin-bottom: 1rem;
-  padding: 1rem 0;
+  margin: 1.5rem 2rem;
 }
 
-.donation-org-row__text {
-  display: flex;
-}
-
-.donation-org-row__text p {
-  padding: 0;
+.medical-needs-heading h1 {
   margin: 0;
 }
 
-.donation-org-row h3 {
-  flex-basis: 40rem;
+.medical-needs-heading .goal-filter {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.text-blue {
+  color: blue;
+}
+
+.package-list {
+  margin: 0 2rem;
+}
+
+.package-card {
+  display: flex;
+  margin-bottom: 2rem;
+  border: 1px solid var(--elixir-color-gray-500);
+}
+
+.card-details {
+  width: 100%;
+  padding: 1rem;
+  display: flex;
+  flex-direction: column;
+}
+
+.package-card .card-details__heading {
+  display: flex;
+  justify-content: space-between;
+  align-items: flex-start;
+  flex: 1;
+}
+
+.package-card .card-details__heading h2 {
+  margin: 0;
+  margin-bottom: 0.5rem;
+}
+
+.package-card .card-details__heading p {
+  margin: 0;
+}
+
+.card_details__package_progress_text {
+  font-size: 1.2rem;
+  font-weight: bold;
+}
+
+.card_details__package_progress_bar {
+  width: 100%;
+  height: 1.5rem;
+  background-color: var(--elixir-color-gray-500);
+  border-radius: 0.5rem;
+}
+
+.card_details__package_progress_bar span {
+  display: block;
+  width: 100%;
+  height: 1.5rem;
+  background-color: var(--elixir-color-red);
+  border-radius: 0.5rem;
 }
 
 @media (max-width: 600px) {
-  .donation-org-row {
-    flex-direction: column;
-  }
-
-  .donation-org-row__text {
-    flex-direction: column;
-  }
-
-  .donation-org-row__text h3 {
-    flex-basis: 0;
+  .package-card img {
+    display: none;
   }
 }
