--- donor-portal/src/pages/donors/styles.css
@@ -1,40 +1,0 @@
-.donors .header-image {
-  color: var(--elixir-color-white);
-}
-
-.donors .header-image::before {
-  content: "";
-  height: 100%;
-  width: 100%;
-  opacity: 30%;
-  position: absolute;
-  z-index: 0;
-}
-
-.donors .header-image h1 {
-  z-index: 1;
-  font-size: 2.5rem;
-  text-align: left;
-  margin-left: 200px;
-}
-
-.donors .text-section {
-  padding-top: 30px;
-  margin: 0 6rem 6rem;
-  line-height: 160%;
-  font-size: 14pt;
-}
-
-.inquiry {
-  text-align: center;
-  font-size: 22px;
-  font-weight: 600;
-  color: var(--elixir-color-red);
-  margin: auto;
-  margin-bottom: 50px;
-}
-
-.inquiry a {
-  text-decoration: underline;
-  color: rgb(35, 35, 146);
-}
