--- donor-portal/src/pages/layout/full-bleed-container/index.tsx
@@ -1,18 +1,0 @@
-import React from "react";
-import "./styles.css";
-
-export default function FullBleedContainer({
-  children,
-  style = {},
-  className,
-}: {
-  children?: React.ReactNode;
-  style?: React.CSSProperties;
-  className?: string;
-}) {
-  return (
-    <div className={`full-bleed-container ${className}`} style={style}>
-      {children}
-    </div>
-  );
-}
