--- donor-portal/src/pages/layout/full-bleed-container/styles.css
@@ -1,8 +1,0 @@
-.full-bleed-container {
-  width: 100vw;
-  position: relative;
-  left: 50%;
-  right: 50%;
-  margin-left: -50vw;
-  margin-right: -50vw;
-}
