+++ donor-portal/src/pages/layout/header-image/styles.css
@@ -1,17 +1,30 @@
-.full-bleed-container {
-  width: 100vw;
+.hero-image {
+  height: 50vh;
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: cover;
   position: relative;
-  left: 50%;
-  right: 50%;
-  margin-left: -50vw;
-  margin-right: -50vw;
 }
 
-.header-image {
-  height: 300px;
-  background-size: cover;
-  background-position: center;
+.hero-image .container {
   display: flex;
-  
-  align-items: center;
+  flex-direction: column;
+  width: 100%;
+  max-width: var(--elixir-content-max-width);
+  margin: 0 auto;
+  color: white;
+}
+
+.hero-image .hero-text {
+  position: absolute;
+  top: 40%;
+}
+
+.hero-image .hero-text .title{
+  font-size: 50px;
+  padding-bottom: 28px;
+}
+
+.hero-image .hero-text .icon{
+  vertical-align: bottom;
 }
