--- donor-portal/src/pages/layout/page/index.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-
-import "./styles.css";
-
-interface PageProps {
-  children: React.ReactNode;
-  className?: string;
-}
-
-export default function Page(props: PageProps) {
-  const { children, className } = props;
-
-  return <main className={`main-container ${className}`}>{children}</main>;
-}
