--- donor-portal/src/pages/layout/page/styles.css
@@ -1,18 +1,0 @@
-:root {
-  --elixir-toc-border-color: var(--elixir-color-gray-300);
-  --elixir-sidebar-width: 300px;
-  --elixir-menu-color: var(--elixir-color-gray-700);
-  --elixir-menu-hover-color: var(--elixir-color-gray-600);
-  --elixir-menu-color-background-hover: var(--elixir-hover-overlay);
-  --elixir-menu-color-background-active: var(--elixir-hover-overlay);
-  --elixir-menu-color-active: var(--elixir-link-color);
-}
-
-.main-container {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  max-width: var(--elixir-content-max-width);
-  margin: 0 auto;
-  min-height: calc(100vh - 145px);
-}
