--- donor-portal/src/pages/login/styles.css
@@ -1,70 +1,0 @@
-.login-container {
-  display: flex;
-  text-align: center;
-  flex-direction: column;
-  margin: 2rem 0;
-  justify-content: center;
-  position: relative;
-}
-
-.login-register-links {
-  display: flex;
-  align-items: center;
-  gap: 2rem;
-  justify-content: center;
-}
-
-.login-container h1 {
-  font-size: 2.5rem;
-  margin: 2rem 0;
-}
-
-.login-container .already-registered {
-  color: var(--elixir-color-gray-700);
-  margin-top: 0;
-}
-
-.login-container .already-registered a {
-  color: var(--elixir-color-gray-700);
-  font-weight: var(--elixir-font-weight-bold);
-}
-
-form {
-  display: flex;
-  flex-direction: column;
-  margin: 0 2rem;
-  align-self: center;
-}
-
-form .input-label {
-  color: var(--elixir-color-gray-700);
-  font-weight: var(--elixir-font-weight-bold);
-  text-align: left;
-  margin: 0.5rem 0 0.2rem;
-  min-width: 500px;
-}
-
-form input {
-  background-color: var(--elixir-color-gray-200);
-  border: 1px solid var(--elixir-color-gray-600);
-  border-radius: 0.2rem;
-  padding: 0.5rem;
-  margin-bottom: 1rem;
-}
-
-form .form-row {
-  display: flex;
-  justify-content: space-between;
-}
-
-@media (max-width: 600px) {
-  form {
-    width: calc(100% - 2rem);
-    margin: 0;
-  }
-
-  form .input-label {
-    min-width: 0;
-    width: 100%;
-  }
-}
