+++ donor-portal/src/pages/newsroom/styles.css
@@ -35,7 +35,6 @@
 
 .news-room .text-section {
   padding-top: 30px;
-  margin: 0 6rem 3rem;
   line-height: 160%;
 }
 
@@ -51,9 +50,8 @@
 
 .news-room .news-article-list {
   display: flex;
-  margin: 3rem 6rem 6rem;
+  margin: 3rem 0;
   flex-direction: column;
-  margin-bottom: 2rem;
 }
 
 .news-room .news-article-list-item {
@@ -62,7 +60,7 @@
   padding-bottom: 1rem;
   margin-bottom: 1rem;
   gap: 0.5rem;
-  border-bottom: var(--elixir-color-gray-900) 1px solid;
+  border-bottom: var(--elixir-color-gray-500) 1px solid;
 }
 
 .news-grid {
